# spec/openapi-spec.yml
openapi: 3.0.3
info:
  title: Regular Income
  description: This is the API for Regular Income
  version: 1.0.0
...
paths:
...
  /api/categories:
    get:
      tags:
        - Category
      summary: Gets all categories flatted
      description: Gets all categories flatted
      operationId: getFlatCategories
      parameters:
        - in: query
          name: deep
          required: false
          schema:
            type: boolean
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      tags:
        - Category
      summary: Creates a new category
      description: Creates a new category
      operationId: createCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryPatch"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
...
  /api/categories/{id}:
    patch:
      tags:
        - Category
      summary: Updates a category
      description: Updates a category
      operationId: updateCategory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryPatch"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      tags:
        - Category
      summary: Deletes a category
      description: Deletes a category
      operationId: deleteCategory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      responses:
        204:
          description: Successful operation
  /api/categories/{parent-id}/children:
    post:
      tags:
        - Category
      summary: Creates a new category as child
      description: Creates a new category as child
      operationId: createCategoryAsChild
      parameters:
        - in: path
          name: parent-id
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryPatch"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /api/categories/{parent-id}/children/{child-id}:
    post:
      tags:
        - Category
      summary: Moves a category to another category as child
      description: Moves a category to another category as child
      operationId: reallocateCategory
      parameters:
        - in: path
          name: parent-id
          required: true
          schema:
            $ref: "#/components/schemas/ID"
        - in: path
          name: child-id
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
...
  /api/reports/balance-progression:
    get:
      tags:
        - Reports
      summary: Get balance progression report
      description: Get balance progression report
      operationId: fetchBalanceProgressionReport
      parameters:
        - in: query
          name: begin
          required: false
          schema:
            $ref: "#/components/schemas/Date"
        - in: query
          name: end
          required: false
          schema:
            $ref: "#/components/schemas/Date"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceProgressionReport"
...
  /api/reports/income-expenses-flow:
    get:
      tags:
        - Reports
      summary: Get income expense flow report
      description: Get income expense flow report
      operationId: fetchIncomeExpenseFlowReport
      parameters:
        - in: query
          name: year
          required: false
          schema:
            type: integer
            format: int32
        - in: query
          name: month
          required: false
          schema:
            type: integer
            format: int32
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeExpenseFlowReport"
  /api/reports//income-expenses-flow/sliding-window/{unit}/{count}:
    get:
      tags:
        - Reports
      summary: Get income expense flow sliding window report
      description: Get income expense flow sliding window report
      operationId: fetchIncomeExpenseFlowSlidingWindowReport
      parameters:
        - in: path
          name: unit
          required: true
          schema:
            $ref: "#/components/schemas/ReportTimeUnits"
        - in: path
          name: count
          required: true
          schema:
            type: integer
            format: int32
        - in: query
          name: reference-date
          required: false
          schema:
            $ref: "#/components/schemas/Date"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeExpenseFlowReport"
components:
  schemas:
    CreatedMetaInformation:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ID"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
    UpdatedMetaInformation:
      type: object
      allOf:
        - $ref: "#/components/schemas/CreatedMetaInformation"
        - properties:
            updatedAt:
              $ref: "#/components/schemas/Timestamp"
    PatchInformation:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ID"
...
    CategoryBudget:
      type: object
      required:
        - budgetInCents
      properties:
        budgetInCents:
          $ref: "#/components/schemas/MoneyValue"
        exceedingThreshold:
          type: number
          format: double
    CategoryPatch:
      type: object
      required:
        - name
      allOf:
        - $ref: "#/components/schemas/PatchInformation"
        - properties:
            name:
              type: string
            description:
              type: string
            budget:
              $ref: "#/components/schemas/CategoryBudget"
    Category:
      type: object
      required:
        - isNew
      allOf:
        - $ref: "#/components/schemas/UpdatedMetaInformation"
        - $ref: "#/components/schemas/CategoryPatch"
        - properties:
            isNew:
              type: boolean
            parentId:
              $ref: "#/components/schemas/ID"
            usageCount:
              type: integer
              format: int64
            subCategories:
              type: array
              items:
                $ref: "#/components/schemas/Category"
...
    BalanceProgressionReport:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/BalanceDataPoint"
    BalanceDataPoint:
      type: object
      properties:
        index:
          type: integer
          format: int32
        date:
          $ref: "#/components/schemas/Date"
        perDayBalanceInCents:
          $ref: "#/components/schemas/MoneyValue"
        balanceInCents:
          $ref: "#/components/schemas/MoneyValue"
    IncomeExpenseFlowReport:
      type: object
      properties:
        start:
          $ref: "#/components/schemas/Date"
        end:
          $ref: "#/components/schemas/Date"
        flows:
          type: array
          items:
            $ref: "#/components/schemas/FlowDataPoint"
    FlowDataPoint:
      type: object
      properties:
        fromCategory:
          type: string
        toCategory:
          type: string
        depthLevel:
          type: integer
          format: int32
        amountInCents:
          $ref: "#/components/schemas/MoneyValue"
...
    ReportTimeUnits:
      type: string
      enum:
        - DAYS
        - WEEKS
        - MONTHS
        - YEARS
        - DECADES
    ID:
      type: string
      format: uuid
    MoneyValue:
      type: integer
      format: int32
...
    Date:
      type: string
      format: date
    Timestamp:
      type: string
      format: date-time
...